use native `useId` from Vue.js 3.5 when available (#3458)
https://github.com/tailwindlabs/headlessui/discussions/3457

Vue 3.5 added a native SSR-safe `useId` function. This PR updates the
code to use that function by default if no custom function has been
provided.

Theoretically this is a breaking change, as versions prior to Vue 3.5
are no longer supported.
However, package managers should emit a warning or an error when the
peer dependency range is no longer satisfied.

---------

Co-authored-by: Robin Malfait <malfait.robin@gmail.com>